microservice/requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
uvicorn
nats-py
